AO-356: Introducing the ActiveObjectsConfigurationFactory.
This separates the concerns of the module descriptor, parsing the configuration, and the factory
creating the configuration.
This factory is also exposed as an OSGi service to be usable in the context of the remote apps
container. This also means exposing the com.atlassian.activeobjects.config package to other
bundles.

Note that this simpifies greatly the (java) module descriptor code and makes it more straight
forward to understand.