AO-357: Make the Admin UI optional.
This change makes all the imported OSGI services used by the admin UI optional.
This means that if any of the services is missing the admin UI will be disabled.
Thanks to a servlet filter we correctly send a 404 in that case, adding extra
logging when in dev mode to help developers understand what's happening.

This is especially useful for the remote app Universal Binary container which
doesn't have a UI.