Make the TransactionSynchronisationManager truly optional.
The cardinality 0..1 is not sufficient to treat a service as optional, one
has to worry about the ServiceNotAvailableException.
Introduce a factory bean that takes care of that for us. Note that the retrieval
of the service is not completely dynamic since it will only be looked up once at
plugin startup. This should be fine since the service is meant to be provided
by the host application.