AO-356: AO scans its configuration from default packages.
Refactored with an ActiveObjects configuration factory which separates
the concerns of the module descriptor, parsing the configuration, and
the factory creating the configuration.

This factory is also used if no module descriptor is defined to create
a configuration from scanned packages: 'ao.model' & 'ao.upgrade'.

Note that this simpifies greatly the (java) module descriptor code and makes it more straight
forward to understand.