AO-387 introduced AOModuleMetaData
ActiveObjectsModuleMetaData contains info about the state of the AO module.
This interface exists so as not to clutter the ActiveObjects api with
excess methods about the AO system itself.
Currently it contains two methods, isInitialized and awaitInitialization but it
is anticipated that methods like, modelVersion and databaseType could be added.
Added in response to code review.